moved computation of scores encoding deadlines from base to assessments module
